# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Jobberwocky",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:251"
        }
    ],
    "paths": {
        "/api/country": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Show all countries for jobs",
                "operationId": "d9d2e793741d06032ade48368011a411",
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/showCountry/{id}": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Displays the details of a country",
                "operationId": "675b944c8d7666b81fc35e8ea305920c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/storeCountry": {
            "post": {
                "tags": [
                    "Country"
                ],
                "summary": "Add a new Country",
                "operationId": "c5bf0e26f39f467bd214e8c9d5505822",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of the new country",
                                        "format": "string",
                                        "default": "Spain"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/updateCountry": {
            "put": {
                "tags": [
                    "Country"
                ],
                "summary": "Upgrade an existing country.",
                "operationId": "1011ed95f3b6ebe680582ccf7a972de6",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of the country",
                                        "format": "number",
                                        "default": "1"
                                    },
                                    "name": {
                                        "description": "name of the country",
                                        "format": "string",
                                        "default": "Spain"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
        "/api/delCountry/{id}": {
            "delete": {
                "tags": [
                    "Country"
                ],
                "summary": "Delete a selected country",
                "operationId": "228f7e06695f17de91e14b513adfc592",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Country identifier",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok."
                    },
                    "404": {
                        "description": "404 Not Found."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}